rename claim extraction to fact checking
