+++ app/repositories/fact-checking.ts
@@ -5,7 +5,8 @@ import { z } from 'zod'
 import { createPost, getPost } from '#app/repositories/post.ts'
 import { type Post } from '#app/types/api-types.ts'
 import { type FactCheck, type DB } from '#app/types/kysely-types.ts'
-import { invariant } from './misc.tsx'
+import { invariant } from '#app/utils/misc.tsx'
+import { MAX_CHARS_PER_POST } from '#app/constants.ts'
 
 export const ClaimExtractionSchema = z.object({
 	claim_context: z
@@ -46,6 +47,10 @@ export const ClaimExtractionSchema = z.object({
 export type ClaimList = z.infer<typeof ClaimExtractionSchema>
 
 export async function extractClaims(content: string): Promise<ClaimList> {
+
+	invariant(content.length <= MAX_CHARS_PER_POST, 'Post content too long')
+	invariant(content.length > 0, 'Post content too short')
+
 	const openai = new OpenAI()
 
 	const completion = await openai.beta.chat.completions.parse({
